Add ASIN field to already published books
Amazon Standard Identification Number (ASIN) is an Amazon-proprietary
10-digit alpha-numeric identifier for Amazon articles. Given the ASIN of
an article, you can create the Amazon URL for this article as follows:

  https://www.amazon.com/dp/<ASIN>

For example:

  https://www.amazon.com/dp/1492057894

The above is the Amazon link for the book "Building Event-Driven
Microservices" which has an ASIN of 1492057894.

Note: for books that have a (deprecated) 10-digit ISBN, the ASIN is
identical to this 10-digit ISBN. However, not all books have a 10-digits
ISBN because 10-digits ISBNs have been deprecated in favour of 13-digits
ISBNs in 2007. For books that have only a 13-digits ISBN, the ASIN is an
independent random string (as for any other types of articles).